requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
langchain
langchain-groq
langchain-community
fastapi
uvicorn
pydantic
streamlit
